refactor: add condition in use cases

+++ src/core/useCases/Expense/DeleteExpense/Rules.ts
@@ -1,3 +1,4 @@
+import { NotFoundError } from "../../../../utils/error";
 import { IExpenseRepository } from "../../../../data/repositories/ExpenseRepository/IExpenseRepository";
 import { DTO } from "./DTO";
 
@@ -6,6 +7,10 @@ export class Rules {
 	constructor(private repository: IExpenseRepository) {}
 
 	async execute({ id }: DTO) {
+
+		const expense = await this.repository.getOne(id);
+
+		if(!expense) throw new NotFoundError("Essa despesa que você quer apagar não existe");
 
 		await this.repository.delete(id);
